Re-worked DataObject so that data is stored in a ByteBuffer, instead of a simple byte array. I took advantage of ByteBuffer access methods to remove a lot of temporary ByteArray and Data stream allocations.

+++ jam/io/hdf/ConvertJamObjToHDFObj.java
@@ -276,7 +276,7 @@ public class ConvertJamObjToHDFObj implements JamHDFFields{
 	 */
 	ScientificDataDimension getSDD(Histogram h, byte numberType) throws HDFException {
 		ScientificDataDimension rval=null;//return value
-		final int rank = h.getDimensionality();
+		final short rank = (short)h.getDimensionality();
 		final int sizeX = h.getSizeX();
 		int sizeY=0;
 		if (rank == 2) {//otherwise rank == 1
